feat: move particle parameters to util
