+++ rust/crates/cpu/src/particles.rs
@@ -7,7 +7,8 @@
 // https://opensource.org/licenses/MIT.
 
 use nalgebra::{Matrix3, Vector3};
-use squishy_volumes_file_frame::{ParticleFlags, ParticleParameters};
+use squishy_volumes_file_frame::ParticleFlags;
+use squishy_volumes_util::ParticleParameters;
 
 #[derive(Default, Debug, Clone)]
 pub struct Particles {
