+++ rust/crates/file_frame/Cargo.toml
@@ -12,3 +12,4 @@ bitflags.workspace = true
 thiserror.workspace = true
 
 squishy_volumes_file_util.path = "../file_util"
+squishy_volumes_util.path = "../util"
