+++ rust/crates/util/Cargo.toml
@@ -14,6 +14,7 @@ rand.workspace = true
 nalgebra.workspace = true
 thiserror.workspace = true
 tracing.workspace = true
+serde.workspace = true
 
 coarse-prof.workspace = true
 coarse-prof.optional = true
